+++ just_audio/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.9.30
 
 * Upgrade ExoPlayer to 2.18.1.
+* Fix bug using headers with LockCachingAudioSource.
+* Add LockCachingAudioSource.resolve().
 
 ## 0.9.29